fix: atualização de review e validação de id pelo controller.

+++ src/reviews/controller.js
@@ -1,10 +1,13 @@
 const { StatusCodes } = require('http-status-codes');
 const reviewSchema = require('./review.schema');
 const service = require('./service');
+const prisma = require('../../prisma/prismaClient');
 
 const create = async (req, res) => {
   const result = reviewSchema.safeParse(req.body);
-  if (!result.success) return res.status(StatusCodes.BAD_REQUEST).json({ error: result.error.format() });
+  if (!result.success) {
+    return res.status(StatusCodes.BAD_REQUEST).json({ error: result.error.format() });
+  }
 
   try {
     const review = await service.createReview(result.data);
@@ -29,9 +32,41 @@ const getById = async (req, res) => {
   }
 };
 
-const remove = async (req, res) => {
+const update = async (req, res) => {
+  const { idUsuario, idFilme } = req.params;
+  const { nota, comentario } = req.body;
+
+  if (req.usuario.tipoUsuario !== 'admin' && req.usuario.id !== Number(idUsuario)) {
+    return res.status(StatusCodes.FORBIDDEN).json({ error: 'Você só pode atualizar suas próprias avaliações.' });
+  }
+
   try {
-    const { idUsuario, idFilme } = req.params;
+    const updatedReview = await prisma.avaliacao.updateMany({
+      where: {
+        idUsuario: Number(idUsuario),
+        idFilme: Number(idFilme),
+      },
+      data: { nota, comentario },
+    });
+
+    if (updatedReview.count === 0) {
+      return res.status(StatusCodes.NOT_FOUND).json({ error: 'Avaliação não encontrada' });
+    }
+
+    res.status(StatusCodes.OK).json({ message: 'Avaliação atualizada com sucesso' });
+  } catch (error) {
+    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ error: error.message });
+  }
+};
+
+const remove = async (req, res) => {
+  const { idUsuario, idFilme } = req.params;
+
+  if (req.usuario.tipoUsuario !== 'admin' && req.usuario.id !== Number(idUsuario)) {
+    return res.status(StatusCodes.FORBIDDEN).json({ error: 'Você só pode deletar suas próprias avaliações.' });
+  }
+
+  try {
     await service.deleteReview(idUsuario, idFilme);
     res.status(StatusCodes.NO_CONTENT).send();
   } catch (err) {
@@ -39,4 +74,4 @@ const remove = async (req, res) => {
   }
 };
 
-module.exports = { create, getAll, getById, remove };
+module.exports = { create, getAll, getById, update, remove };
